# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Astervia_wacraft-core_src_campaign_entity.Campaign:
    properties:
      created_at:
        type: string
      id:
        type: string
      messaging_product:
        allOf:
        - $ref: '#/definitions/github_com_Astervia_wacraft-core_src_messaging-product_entity.MessagingProduct'
        description: Null if manager received the message.
      messaging_product_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      workspace:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_entity.Workspace'
      workspace_id:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_common_model.DescriptiveError:
    properties:
      content:
        description: Additional information about the error provided by the service
          that generated it. E. g. if an ORM returns an error, it will be propagated
          here as the raw error.
      message:
        description: Short text describing the error as best as we can for humans.
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_common_model.RequiredID:
    properties:
      id:
        description: The unique identifier.
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_contact_entity.Contact:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      photo_path:
        type: string
      updated_at:
        type: string
      workspace:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_entity.Workspace'
      workspace_id:
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_contact_model.CreateContact:
    properties:
      email:
        type: string
      name:
        type: string
      photo_path:
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_contact_model.UpdateContact:
    properties:
      email:
        type: string
      id:
        description: The unique identifier.
        type: string
      name:
        type: string
      photo_path:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_messaging-product_entity.MessagingProduct:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        allOf:
        - $ref: '#/definitions/github_com_Astervia_wacraft-core_src_messaging-product_model.MessagingProductName'
        description: Add type:enum('WhatsApp'); when it becomes supported by GORM
          and PostgreSQL
        enum:
        - WhatsApp
      phone_config:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig'
      phone_config_id:
        type: string
      updated_at:
        type: string
      workspace:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_entity.Workspace'
      workspace_id:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_messaging-product_model.MessagingProductName:
    enum:
    - WhatsApp
    type: string
    x-enum-varnames:
    - WhatsApp
...
  github_com_Astervia_wacraft-core_src_phone-config_entity.PhoneConfig:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      display_phone:
        description: Display phone number (e.g., +1234567890)
        type: string
      id:
        type: string
      is_active:
        type: boolean
      meta_app_secret:
        type: string
      name:
        description: Friendly name for this configuration
        type: string
      updated_at:
        type: string
      waba_account_id:
        description: WhatsApp Business Account ID
        type: string
      waba_id:
        description: Phone Number ID from Meta (unique when active)
        type: string
      webhook_verify_token:
        type: string
      workspace:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_entity.Workspace'
      workspace_id:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_user_entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: 'Deprecated: use workspace policies'
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_workspace_entity.Workspace:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/github_com_Astervia_wacraft-core_src_user_entity.User'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: wacraft@astervia.tech
    name: Astervia Dev Team
    url: https://github.com/Astervia
  description: Backend server for the wacraft project. Handles WhatsApp Cloud API
    operations, including message sending, receiving, and webhook handling.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wacraft Server API
  version: 0.1.0
paths:
...
  /campaign/content/{keyName}/like/{likeText}:
    get:
      consumes:
      - application/json
      description: Applies a ILIKE filter on the specified key field and returns paginated
        results.
      parameters:
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - default: 10
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - in: query
        name: messaging_product_id
        type: string
      - in: query
        name: name
        type: string
      - default: 0
        description: The offset from where to start the items
        in: query
        name: offset
        type: integer
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      - in: query
        name: workspace_id
        type: string
      - in: path
        name: keyName
        required: true
        type: string
      - in: path
        maxLength: 512
        minLength: 1
        name: likeText
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of matching campaigns
          schema:
            items:
              $ref: '#/definitions/github_com_Astervia_wacraft-core_src_campaign_entity.Campaign'
            type: array
        "400":
          description: Invalid input (e.g., decoding or query error)
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Search campaigns with regex-like operator
      tags:
      - Campaign
...
  /contact:
    delete:
      consumes:
      - application/json
      description: Deletes a contact based on the ID sent in the request body.
      parameters:
      - description: Contact ID to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.RequiredID'
      produces:
      - application/json
      responses:
        "204":
          description: Contact deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Delete contact by ID
      tags:
      - Contact
    get:
      consumes:
      - application/json
      description: Returns a paginated list of contacts using optional query parameters
        for filtering and sorting.
      parameters:
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - in: query
        name: email
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - default: 10
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - in: query
        name: name
        type: string
      - default: 0
        description: The offset from where to start the items
        in: query
        name: offset
        type: integer
      - in: query
        name: photo_path
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      - in: query
        name: workspace_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of contacts
          schema:
            items:
              $ref: '#/definitions/github_com_Astervia_wacraft-core_src_contact_entity.Contact'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Get contacts paginated
      tags:
      - Contact
    post:
      consumes:
      - application/json
      description: Creates a new contact using the provided data and returns the created
        object.
      parameters:
      - description: Contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_contact_model.CreateContact'
      produces:
      - application/json
      responses:
        "201":
          description: Created contact
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_contact_entity.Contact'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Create a new contact
      tags:
      - Contact
    put:
      consumes:
      - application/json
      description: Updates an existing contact using the provided ID and data.
      parameters:
      - description: Updated contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_contact_model.UpdateContact'
      produces:
      - application/json
      responses:
        "200":
          description: Updated contact
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_contact_entity.Contact'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Update contact
      tags:
      - Contact
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"